Adds RaiseTask class to assert API responses for failing tasks. Closes #2579
